Fix home page search capability * If you deleted an item, the item would not actually delete. This was due to setting the sessionStorage arrays to the "popped" or spliced out items. If you had [1,2,3], we'd perform splice(0,deleteCount: 1) and the array would store this as [1] * Added "enter" key functionality since typeahead kinda pushes for this "enter" feeling. * If you looked up the multi-organization search then went back to home (assuming all items are stilled filled in including the typeahead value), if you backspaced your item wouldn't be deleted from sessionStorage. That is now fixed.
In general, this was an issue if you had done a selectOne, i.e. click on a suggestion; the selectOne adds your item to an array in which the client browser performs the multi-org search. If you were to backspace in the typeahead field (assuming the field's item was in the data array), then the item would still be in the array. So this has been fixed.

Misc.
* The base.html shows the author and description of the site now
* multiple_orgs.html doesn't show 1 Item(s) thing. The "s" is only appended if the number of items is truly not "1"